Move chisel/... to chisel/core/..., make chisel/compatibility package/folder, move more things into utils
