+++ chiselFrontend/src/main/scala/chisel/core/Aggregate.scala
