+++ chiselFrontend/src/main/scala/chisel/core/Bits.scala
