+++ chiselFrontend/src/main/scala/chisel/core/Module.scala
