+++ chiselFrontend/src/main/scala/chisel/core/Reg.scala
