+++ chiselFrontend/src/main/scala/chisel/core/SeqUtils.scala
